+++ include/drjit/packet_sse42.h
@@ -210,7 +210,7 @@ template <bool IsMask_, typename Derived_> struct alignas(16)
     DRJIT_INLINE auto eq_ (Ref a) const {
         using Int = int_array_t<Derived>;
         if constexpr (IsMask_)
-            return mask_t<Derived>(eq(Int(derived()), Int(a)));
+            return mask_t<Derived>(Int(derived()) == Int(a));
         else
             return DRJIT_COMP(eq, EQ_OQ);
     }
@@ -218,7 +218,7 @@ template <bool IsMask_, typename Derived_> struct alignas(16)
     DRJIT_INLINE auto neq_(Ref a) const {
         using Int = int_array_t<Derived>;
         if constexpr (IsMask_)
-            return mask_t<Derived>(neq(Int(derived()), Int(a)));
+            return mask_t<Derived>(Int(derived()) != Int(a));
         else
             return DRJIT_COMP(neq, NEQ_UQ);
     }
@@ -1027,7 +1027,7 @@ template <bool IsMask_, typename Derived_> struct alignas(16)
     DRJIT_INLINE auto eq_ (Ref a) const {
         using Int = int_array_t<Derived>;
         if constexpr (IsMask_)
-            return mask_t<Derived>(eq(Int(derived()), Int(a)));
+            return mask_t<Derived>(Int(derived()) == Int(a));
         else
             return DRJIT_COMP(eq, EQ_OQ);
     }
@@ -1035,7 +1035,7 @@ template <bool IsMask_, typename Derived_> struct alignas(16)
     DRJIT_INLINE auto neq_(Ref a) const {
         using Int = int_array_t<Derived>;
         if constexpr (IsMask_)
-            return mask_t<Derived>(neq(Int(derived()), Int(a)));
+            return mask_t<Derived>(Int(derived()) != Int(a));
         else
             return DRJIT_COMP(neq, NEQ_UQ);
     }